+++ src/lib/components/navigation/breadcrumb/index.ts
@@ -2,3 +2,4 @@ export {default as Container} from "./BreadcrumbContainer.svelte";
 
 export {default as Anchor} from "./BreadcrumbAnchor.svelte";
 export {default as Item} from "./BreadcrumbItem.svelte";
+export {default as Group} from "./BreadcrumbGroup.svelte";